requirements.txt
telethon>=1.34.0
openai>=1.0.0
python-dotenv
cryptg
qrcode
aiohttp>=3.9.0
